Added atalk example and spent a while fixing the memory prefilling code

+++ src/pub_sec_mutations.rs
@@ -51,7 +51,7 @@ where
             _ => panic!("Should have current mutate target set to secret")
         };
 
-        let sample = state.rand_mut().next();
+        let sample = state.rand_mut().next() as u32;
         input.set_secret_part_bytes(&sample.to_ne_bytes());
         Ok(MutationResult::Mutated)
     }
